@@ -15,30 +15,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.8",
-    "@angular/common": "^5.2.8",
-    "@angular/compiler": "^5.2.8",
-    "@angular/core": "^5.2.8",
-    "@angular/forms": "^5.2.8",
-    "@angular/http": "^5.2.8",
-    "@angular/platform-browser": "^5.2.8",
-    "@angular/platform-browser-dynamic": "^5.2.8",
-    "@angular/router": "^5.2.8",
-    "core-js": "^2.5.3",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.20"
+    "@angular/animations": "6.0.0",
+    "@angular/common": "6.0.0",
+    "@angular/compiler": "6.0.0",
+    "@angular/core": "6.0.0",
+    "@angular/forms": "6.0.0",
+    "@angular/http": "6.0.0",
+    "@angular/platform-browser": "6.0.0",
+    "@angular/platform-browser-dynamic": "6.0.0",
+    "@angular/router": "6.0.0",
+    "core-js": "^2.5.5",
+    "rxjs": "6.1.0",
+    "zone.js": "0.8.26",
+    "rxjs-compat": "^6.1.0"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.3",
-    "@angular/compiler-cli": "^5.2.8",
+    "@angular/cli": "^6.0.0",
+    "@angular/compiler-cli": "6.0.0",
     "@types/node": "~6.0.60",
-    "codelyzer": "^4.2.1",
-    "jest": "^22.4.2",
-    "jest-preset-angular": "^5.2.1",
+    "codelyzer": "^4.3.0",
+    "jest": "^22.4.3",
+    "jest-preset-angular": "^5.2.2",
     "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "2.6.2"
+    "ts-node": "^6.0.2",
+    "tslint": "^5.10.0",
+    "typescript": "2.7.2"
   },
   "jest": {
     "preset": "jest-preset-angular",
