@@ -1,5 +1,6 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
+
 const image = require('assets/its_something.png');
 
 @Component({
