@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core'
 import { HttpClient, HttpErrorResponse, HttpParams } from '@angular/common/http'
-
-import { Observable } from 'rxjs/Observable'
-import { of } from 'rxjs/observable/of'
+import { Observable } from 'rxjs'
+import { of } from 'rxjs'
 import { catchError } from 'rxjs/operators'
+
 import { Hero } from './hero'
 
 export const heroesUrl = 'api/heroes' // URL to web api
